requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
gradio>=4.0.0
openai-whisper>=20231117
